Chore: move files into organized folders
