--- src/utils.tsx
@@ -1,2 +1,0 @@
-
-export {}
